Refactor to use org.geoserver.geopkg consistently
