+++ src/extension/geopkg-output/src/main/java/org/geoserver/geopkg/GeoPkgDataStoreFactoryInitializer.java
@@ -3,7 +3,7 @@
  * This code is licensed under the GPL 2.0 license, available at the root
  * application directory.
  */
-package org.geoserver.geopackage;
+package org.geoserver.geopkg;
 
 import org.geoserver.data.DataStoreFactoryInitializer;
 import org.geoserver.platform.GeoServerResourceLoader;
